+++ commons/src/test/java/com/karuslabs/commons/util/PointTest.java
@@ -31,17 +31,13 @@ import org.bukkit.Location;
 import org.bukkit.util.Vector;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.*;
-
-import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.params.provider.Arguments.of;
 
 
-@ExtendWith(MockitoExtension.class)
 class PointTest {
     
     static final Point POINT = new Point(null, 1, 2, 3, 90, 180);