+++ commons/src/test/java/com/karuslabs/commons/util/collection/ConcurrentClassMapTest.java
@@ -21,19 +21,15 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-
 package com.karuslabs.commons.util.collection;
 
 import java.util.concurrent.ConcurrentHashMap;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 
-@ExtendWith(MockitoExtension.class)
 class ConcurrentClassMapTest {
     
     @Test
@@ -45,7 +41,6 @@ class ConcurrentClassMapTest {
 }
 
 
-@ExtendWith(MockitoExtension.class)
 class ConcurrentHashClassMapTest {
     
     @Test
@@ -57,7 +52,6 @@ class ConcurrentHashClassMapTest {
 }
 
 
-@ExtendWith(MockitoExtension.class)
 class ConcurrentProxiedClassMapTest {
     
     @Test