+++ commons/src/test/java/com/karuslabs/commons/util/collection/ConcurrentTokenMapTest.java
@@ -21,7 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-
 package com.karuslabs.commons.util.collection;
 
 import com.karuslabs.commons.util.collection.TokenMap.Key;
@@ -30,20 +29,16 @@ import java.util.concurrent.*;
 import java.util.stream.Stream;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.*;
-
-import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 
-@ExtendWith(MockitoExtension.class)
 class ConcurrentTokenMapTest {
     
     @ParameterizedTest
-    @MethodSource({"parameters"})
+    @MethodSource({"maps"})
     void containsKey(ConcurrentTokenMap<String, Object> map) {
         map.put("a", String.class, "b");
         assertTrue(map.containsKey("a", String.class));
@@ -51,7 +46,7 @@ class ConcurrentTokenMapTest {
     
     
     @ParameterizedTest
-    @MethodSource({"parameters"})
+    @MethodSource({"maps"})
     void get(ConcurrentTokenMap<String, Object> map) {
         map.put("a", int.class, 1);
         assertEquals(1, (int) map.get("a", int.class));
@@ -59,7 +54,7 @@ class ConcurrentTokenMapTest {
     
     
     @ParameterizedTest
-    @MethodSource({"parameters"})
+    @MethodSource({"maps"})
     void getOrDefault(ConcurrentTokenMap<String, Object> map) {
         map.map().put(TokenMap.key("a", int.class), map);
         assertEquals(1, (int) map.getOrDefault("a", int.class, 1));
@@ -67,7 +62,7 @@ class ConcurrentTokenMapTest {
     
     
     @ParameterizedTest
-    @MethodSource({"parameters"})
+    @MethodSource({"maps"})
     void remove(ConcurrentTokenMap<String, Object> map) {
         map.put(TokenMap.key("a", int.class), 1);
         map.remove("a", int.class);
@@ -76,7 +71,7 @@ class ConcurrentTokenMapTest {
     }
     
     
-    static Stream<ConcurrentTokenMap<String, Object>> parameters() {
+    static Stream<ConcurrentTokenMap<String, Object>> maps() {
         ConcurrentTokenMap<String, Object> hashed = ConcurrentTokenMap.of();
         ConcurrentTokenMap<String, Object> proxied = ConcurrentTokenMap.of(new ConcurrentHashMap<>());
         
@@ -86,7 +81,6 @@ class ConcurrentTokenMapTest {
 }
 
 
-@ExtendWith(MockitoExtension.class)
 class ConcurrentHashTokenMapTest {
     
     ConcurrentTokenMap<String, Object> map = ConcurrentTokenMap.of(1);
@@ -100,7 +94,6 @@ class ConcurrentHashTokenMapTest {
 }
 
 
-@ExtendWith(MockitoExtension.class)
 class ConcurrentProxiedTokenMapTest {
     
     ConcurrentMap<Key<String, ? extends Object>, Object> proxied = new ConcurrentHashMap<>();