+++ commons/src/test/java/com/karuslabs/commons/util/collection/TrieSetTest.java
@@ -24,13 +24,10 @@
 package com.karuslabs.commons.util.collection;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 
-@ExtendWith(MockitoExtension.class)
 class TrieSetTest {
     
     TrieSet set = new TrieSet();