+++ commons/src/test/java/com/karuslabs/commons/util/collection/TrieTest.java
@@ -28,16 +28,12 @@ import com.karuslabs.commons.util.collection.Trie.TrieIterator;
 import java.util.*;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.*;
-
-import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 
-@ExtendWith(MockitoExtension.class)
 class TrieTest {
     
     static final Trie<String> TRIE = new Trie<>();