+++ commons/src/test/java/com/karuslabs/commons/util/concurrent/locks/AutoReadWriteLockTest.java
@@ -47,7 +47,7 @@ import static org.mockito.quality.Strictness.LENIENT;
 class AutoReadWriteLockTest {
     
     @ParameterizedTest
-    @MethodSource({"lock_parameters"})
+    @MethodSource({"locks"})
     void lock(Lock wrapper, Lock lock) throws InterruptedException {
         wrapper.lock();
         verify(lock).lock();
@@ -72,7 +72,7 @@ class AutoReadWriteLockTest {
 
     
     @ParameterizedTest
-    @MethodSource({"lock_parameters"})
+    @MethodSource({"locks"})
     void hold(Holdable wrapper, Lock lock) throws InterruptedException {
         try (var mutex = wrapper.hold()) {
             verify(lock).lock();
@@ -83,7 +83,7 @@ class AutoReadWriteLockTest {
     }
     
     @ParameterizedTest
-    @MethodSource({"lock_parameters"})
+    @MethodSource({"locks"})
     void holdInterruptibly(Holdable wrapper, Lock lock) throws InterruptedException {
         try (var mutex = wrapper.holdInterruptibly()) {
             verify(lock).lockInterruptibly();
@@ -94,7 +94,7 @@ class AutoReadWriteLockTest {
     } 
     
     
-    static Stream<Arguments> lock_parameters() {
+    static Stream<Arguments> locks() {
         ReadLock reader = when(mock(ReadLock.class).toString()).thenReturn("delegate").getMock();
         var autoreader = new AutoReadLock(new AutoReadWriteLock(), reader);